@@ -1,23 +1,78 @@
-﻿using System.IO;
-using Microsoft.AspNetCore;
-using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.Authorization;
+using Serilog;
 
-namespace CalculatorApi;
+Log.Logger = new LoggerConfiguration()
+    .WriteTo.Console()
+    .CreateBootstrapLogger();
 
-public class Program
+Log.Information("Starting up");
+
+try
 {
-    public static void Main(string[] args)
+    var builder = WebApplication.CreateBuilder(args);
+
+    // Add Serilog to the container.
+    builder.Host.UseSerilog((hostingContext, loggerConfiguration) => loggerConfiguration
+        .ReadFrom.Configuration(hostingContext.Configuration));
+
+    // Add services to the container.
+
+    builder.Services.AddControllers();
+    // Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
+    builder.Services.AddEndpointsApiExplorer();
+    builder.Services.AddSwaggerGen();
+
+    builder.Services.AddMetrics();
+
+    var authority = builder.Configuration["Auth:Authority"];
+    var audience = builder.Configuration["Auth:Audience"];
+
+    builder.Services.AddAuthentication("Bearer")
+        .AddJwtBearer("Bearer", options =>
+        {
+            options.Authority = authority;
+            options.Audience = audience;
+            options.TokenValidationParameters.ValidateAudience = false;
+            options.TokenValidationParameters.ValidTypes = new[] { "JWT" };
+            options.TokenValidationParameters.ValidateLifetime = true;
+        });
+
+    builder.Services.AddAuthorization(options =>
     {
-        CreateWebHostBuilder(args).Build().Run();
+            options.AddPolicy("calc:double",
+                policy => policy.Requirements.Add(new HasScopeRequirement("calc:double", authority)));
+            options.AddPolicy("calc:square",
+                policy => policy.Requirements.Add(new HasScopeRequirement("calc:square", authority)));
+    });
+
+    // Register the scope authorization handler
+    builder.Services.AddSingleton<IAuthorizationHandler, HasScopeHandler>();
+
+    var app = builder.Build();
+
+    // Configure the HTTP request pipeline.
+    if (app.Environment.IsDevelopment())
+    {
+        app.UseSwagger();
+        app.UseSwaggerUI();
     }
 
-    public static IWebHostBuilder CreateWebHostBuilder(string[] args)
-    {
-        return WebHost.CreateDefaultBuilder(args)
-            .UseUrls("http://*:5001")
-            .UseKestrel()
-            .UseContentRoot(Directory.GetCurrentDirectory())
-            .UseIISIntegration()
-            .UseStartup<Startup>();
-    }
+    app.UseHttpsRedirection();
+
+    app.UseAuthorization();
+    app.UseAuthorization();
+
+    app.MapControllers().RequireAuthorization();
+    app.MapControllers();
+
+    app.Run();
+}
+catch (Exception ex)
+{
+    Log.Fatal(ex, "Unhandled exception");
+}
+finally
+{
+    Log.Information("Shut down complete");
+    Log.CloseAndFlush();
 }
